+++ doc/src/base/arrays.md
@@ -188,30 +188,3 @@ and can be converted to/from the latter via `Array(bitarray)` and `BitArray(arra
 ```@docs
 Base.flipbits!
 ```
-
-## [Sparse Vectors and Matrices](@id stdlib-sparse-arrays)
-
-Sparse vectors and matrices largely support the same set of operations as their dense counterparts.
-The following functions are specific to sparse arrays.
-
-```@docs
-Base.SparseArrays.SparseVector
-Base.SparseArrays.SparseMatrixCSC
-Base.SparseArrays.sparse
-Base.SparseArrays.sparsevec
-Base.SparseArrays.issparse
-Base.SparseArrays.nnz
-Base.SparseArrays.spzeros
-Base.SparseArrays.spdiagm
-Base.SparseArrays.sprand
-Base.SparseArrays.sprandn
-Base.SparseArrays.nonzeros
-Base.SparseArrays.rowvals
-Base.SparseArrays.nzrange
-Base.SparseArrays.dropzeros!(::SparseMatrixCSC, ::Bool)
-Base.SparseArrays.dropzeros(::SparseMatrixCSC, ::Bool)
-Base.SparseArrays.dropzeros!(::SparseVector, ::Bool)
-Base.SparseArrays.dropzeros(::SparseVector, ::Bool)
-Base.SparseArrays.permute
-Base.permute!{Tv, Ti, Tp <: Integer, Tq <: Integer}(::SparseMatrixCSC{Tv,Ti}, ::SparseMatrixCSC{Tv,Ti}, ::AbstractArray{Tp,1}, ::AbstractArray{Tq,1})
-```
