+++ stdlib/SparseArrays/src/SparseArrays.jl
@@ -1,5 +1,7 @@
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
+__precompile__(true)
+
 """
 Support for sparse arrays. Provides `AbstractSparseArray` and subtypes.
 """
@@ -36,5 +38,14 @@ include("sparsematrix.jl")
 include("sparsevector.jl")
 include("higherorderfns.jl")
 include("linalg.jl")
+include("deprecated.jl")
+
+
+# temporarily moved here and commented out from from base/linalg/diagonal.jl, base/linalg/tridiag.jl
+# and base/linalg/bidiag.jl due to their usage of spzeros
+similar(B::Bidiagonal, ::Type{T}, dims::Union{Dims{1},Dims{2}}) where {T} = spzeros(T, dims...)
+similar(D::Diagonal, ::Type{T}, dims::Union{Dims{1},Dims{2}}) where {T} = spzeros(T, dims...)
+similar(S::SymTridiagonal, ::Type{T}, dims::Union{Dims{1},Dims{2}}) where {T} = spzeros(T, dims...)
+similar(M::Tridiagonal, ::Type{T}, dims::Union{Dims{1},Dims{2}}) where {T} = spzeros(T, dims...)
 
 end
