+++ stdlib/SparseArrays/test/runtests.jl
@@ -1,0 +1,7 @@
+# This file is a part of Julia. License is MIT: https://julialang.org/license
+
+using Test, SparseArrays
+
+include("higherorderfns.jl")
+include("sparse.jl")
+include("sparsevector.jl")
