+++ stdlib/SuiteSparse/test/runtests.jl
@@ -1,7 +1,7 @@
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
 using Test
-using SuiteSparse
+using SuiteSparse, SparseArrays
 
 if Base.USE_GPL_LIBS
    include("umfpack.jl")
