+++ test/arrayops.jl
@@ -3,6 +3,7 @@
 # Array test
 isdefined(Main, :TestHelpers) || @eval Main include("TestHelpers.jl")
 using Main.TestHelpers.OAs
+using SparseArrays
 
 @testset "basics" begin
     @test length([1, 2, 3]) == 3
