+++ test/linalg/special.jl
@@ -1,6 +1,7 @@
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
 using Test
+using SparseArrays
 
 using Base.LinAlg: mul!
